refactor: split codegen orchestration
